requirements.txt
Django>=4.0,<5.0
gunicorn>=20.1.0
mysqlclient>=2.0.3
djangorestframework>=3.13.1
Pillow>=8.4.0
whitenoise>=6.0.0